Add files via upload
A deep fake can replace a person in an existing recording with a similar person.
While most the people are using this as entertainment, others could be harmful to
individuals and society. For example, Misuse of your bank account information
by phone. Since the malicious use of deep fakes is dangerous, it is important to
recognize such attacks and distinguish between real and fake ones.
In this project, an attempt is made to detect fake audio files from the real ones.
To reach this aim, we use a visual representation of audio signals. One of the
most important visual representations of audio files called spectrograms. Spectrograms and Constant-Q-transform can be used as a way of visualizing the
change of a nonstationary signal’s frequency content over time. After the feature
extraction phase, an attempt is made to use learning algorithms such as logistic
regression, support vector machine, Naive base, and deep neural network to
detect fake audios. Finally, the efficiency of the proposed approach is tested on
the Fake-or-Real dataset.